requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
python-docx>=1.1.0
